Add some policies to check the validty of the selector, add some services.
Add some policies classes and interfaces to be checked against using the
selector string to make it respect some constraints.

Add a service to check the validity of a selector string using such policies.

Add a service to broadcast new selector string updates.

Add some template code in HTML files to check that this system works.

Minor fixes and updates.